fix: parameter handling robustness and thread limit
- Thread count changed from u8 to usize, removing the 255-thread
  ceiling and the lossy i64 cast in the clap range validator.

- pipeline encoder fn: replaced unwrap() calls on clap get_one with
  unwrap_or fallbacks for quality, colorspace, alpha_mode, speed,
  and slight_loss parameters. Makes the encoder function safe to
  call without CLI defaults.

- ResizeValue from_str: compiled regex patterns are now stored in
  LazyLock statics rather than recompiled on every parse.

- Fixed typo "chrome quality" -> "chroma quality" in mozjpeg CLI
  help text.

+++ src/cli/common.rs
@@ -38,7 +38,7 @@ impl CommonArgs for Command {
 
                 Usage of multiple threads can speed up the execution of tasks, especially on multi-core processors.
                 By default, the number of available threads is utilized."#})
-                .value_parser(value_parser!(u8).range(1..=threads::num_threads() as i64)),
+                .value_parser(value_parser!(usize).range(1..=threads::num_threads())),
             arg!(-x --strip "Strip metadata when encoding images (where supported)")
                 .action(ArgAction::SetTrue),
             arg!(--"no-progress" "Disables progress bar.")
